Clarify port and protocol requirements
The existing doc makes it sound like I need to allow traffic to the listed ports inbound on my firewall in addition to outbound. In reality I need only allow outbound to those destination ports, along with return traffic, right?

The chart also lists ephemeral ports for TURN/STUN, which I think adds to the confusion.

My change re-words the directionality requirement to simply say "outbound" (allowing corresponding return traffic is assumed, I think, in the age of stateful firewalls). It also removes the high-port range for TURN/STUN from the chart, which aligns that chart to a similar chart in "Amazon Connect Troubleshooting and Best Practices"->"Network Ports and Protocols" of the Administrator Guide. After this change, one discrepency remains between the charts in the two docs: the presence of port 80 in this chart.